pyproject.toml
[project]
name = "godmode"
version = "0.1.0"
description = "Numeric market-state recorder (research engine) per SPEC.md v1.1.1"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Proprietary" }
authors = [{ name = "GodMode" }]

dependencies = [
  "numpy>=1.26",
  "pandas>=2.2",
  "pyarrow>=15",
  "pydantic>=2.7",
  "pyyaml>=6.0",
  "websockets>=12.0",
  "click>=8.1",
  "httpx>=0.27",
  "fastapi>=0.110",
  "uvicorn>=0.30",
  "jinja2>=3.1",
  "python-multipart>=0.0.9",
  "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "ruff>=0.5",
]

[project.scripts]
godmode = "godmode.cli.main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]



